+++ packages/stream_video_flutter/lib/src/call_screen/call_content/picture_in_picture/android_pip_overlay.dart
@@ -57,6 +57,7 @@ class _AndroidPipOverlayState extends State<AndroidPipOverlay>
 
     _participantsSubscription = widget.call
         .partialState((state) => state.callParticipants)
+        .throttleByCollectionSize()
         .listen(recalculateParticipants);
   }